+++ packages/transaction-pay-controller/src/index.ts
@@ -7,6 +7,7 @@ export type {
   TransactionPayControllerMessenger,
   TransactionPayControllerOptions,
   TransactionPayControllerSetIsMaxAmountAction,
+  TransactionPayControllerSetIsPostQuoteAction,
   TransactionPayControllerState,
   TransactionPayControllerStateChangeEvent,
   TransactionPayControllerUpdatePaymentTokenAction,