+++ packages/transaction-pay-controller/src/utils/token.test.ts
@@ -471,4 +471,5 @@ describe('Token Utils', () => {
       ]);
     });
   });
+
 });